Add list view for shortened URLs
This replaces the previous behavior, showing a 404 on visitng the root
of a Lessn install, with a list of the short URLs that have been created

As a result of needing them in more places, the constants used
throughout Lessn have been broken out to their own file, and a new one
has been added: `LESSN_ROOT`, which is used by the new list view code.

Instructions regarding viewing the new list have been added to
the README. These clarify that you need to visit the control panel to
initialize Lessn.